refactor (LottoNumberPool) : data package 로 옮기고, LottoNumber 의 of의 접근지시자를 protect 로 변경
외부 패키지에서 LottoNumberPool만 사용할 수 있도록 하기 위함
