+++ src/main/java/lotto/domain/data/LottoNumberPool.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.data;
 
 import lotto.domain.data.LottoNumber;
